# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Speech Analysis API",
...
    "contact": {
      "name": "Speech Analysis API Support",
      "url": "https://github.com/your-username/audio-analysis",
      "email": "support@yourapi.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0",
    "x-logo": {
      "url": "https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png",
      "altText": "Speech Analysis API"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.yourapi.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/analyze/pronunciation": {
      "post": {
        "summary": "Pronunciation",
        "description": "Pronunciation analysis using audio-to-phoneme recognition (no text transcription needed).",
        "operationId": "pronunciation_analyze_pronunciation_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_pronunciation_analyze_pronunciation_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/keys": {
      "get": {
        "tags": [
          "Admin - API Keys"
        ],
        "summary": "List API Keys",
        "description": "Retrieve all API keys with usage statistics and limits",
        "operationId": "get_api_keys_api_admin_keys_get",
        "responses": {
          "200": {
            "description": "List of API keys retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeysListResponse"
                },
                "example": {
                  "api_keys": [
                    {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "description": "Production API Key",
                      "is_active": true,
                      "usage_count": 150,
                      "minute_usage": 2,
                      "daily_usage": 45,
                      "monthly_usage": 150,
                      "minute_limit": 60,
                      "daily_limit": 1000,
                      "monthly_limit": 10000,
                      "last_used_at": "2024-01-15T10:30:00Z",
                      "created_at": "2024-01-01T09:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin - API Keys"
        ],
        "summary": "Create API Key",
        "description": "Create a new API key with custom rate limits.\n    \n    **⚠️ Important**: The API key is only shown once for security reasons. \n    Make sure to copy and store it securely.",
        "operationId": "create_api_key_api_admin_keys_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_api_key_api_admin_keys_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreateResponse"
                },
                "example": {
                  "api_key": "sk-1234567890abcdef1234567890abcdef1234567890abcdef",
                  "key_id": "123e4567-e89b-12d3-a456-426614174000",
                  "description": "Production API Key",
                  "minute_limit": 60,
                  "daily_limit": 1000,
                  "monthly_limit": 10000
                }
              }
            }
          },
          "400": {
            "description": "Invalid input parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/keys/{key_id}": {
      "put": {
        "tags": [
          "Admin - API Keys"
        ],
        "summary": "Update API Key",
        "description": "Update API key settings such as rate limits, description, or active status",
        "operationId": "update_api_key_api_admin_keys__key_id__put",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/APIKeyUpdateRequest"
                  }
                ],
                "description": "Fields to update",
                "title": "Updates"
              },
              "example": {
                "description": "Updated Production Key",
                "is_active": true,
                "minute_limit": 120,
                "daily_limit": 2000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key updated successfully",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "success": true
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to update API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/analytics": {
      "get": {
        "tags": [
          "Admin - Analytics"
        ],
        "summary": "Usage Analytics",
        "description": "Get comprehensive usage analytics and statistics for all API keys",
        "operationId": "get_usage_analytics_api_admin_analytics_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageAnalyticsResponse"
                },
                "example": {
                  "api_keys": [
                    {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "total_requests": 1500,
                      "last_request_at": "2024-01-15T10:30:00Z",
                      "first_request_at": "2024-01-01T09:00:00Z"
                    }
                  ],
                  "recent_logs": [
                    {
                      "endpoint": "/analyze/pronunciation",
                      "deep_analysis": true,
                      "use_audio": true,
                      "created_at": "2024-01-15T10:30:00Z"
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "APIKeyCreateResponse": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "key_id": {
            "type": "string",
            "title": "Key Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "minute_limit": {
            "type": "integer",
            "title": "Minute Limit"
          },
          "daily_limit": {
            "type": "integer",
            "title": "Daily Limit"
          },
          "monthly_limit": {
            "type": "integer",
            "title": "Monthly Limit"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "key_id",
          "description",
          "minute_limit",
          "daily_limit",
          "monthly_limit"
        ],
        "title": "APIKeyCreateResponse",
        "description": "Response model for API key creation"
      },
...
      "APIKeyUpdateRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "minute_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minute Limit"
          },
          "daily_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Limit"
          },
          "monthly_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Limit"
          }
        },
        "type": "object",
        "title": "APIKeyUpdateRequest",
        "description": "Request model for updating API keys"
      },
      "APIKeysListResponse": {
        "properties": {
          "api_keys": {
            "items": {
              "$ref": "#/components/schemas/APIKeyInfo"
            },
            "type": "array",
            "title": "Api Keys"
          }
        },
        "type": "object",
        "required": [
          "api_keys"
        ],
        "title": "APIKeysListResponse",
        "description": "Response model for listing API keys"
      },
      "AnalyzeResponse": {
        "properties": {
          "pronunciation": {
            "$ref": "#/components/schemas/PronunciationResult"
          },
          "predicted_text": {
            "type": "string",
            "title": "Predicted Text"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpeechMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "grammar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GrammarCorrection"
              },
              {
                "type": "null"
              }
            ]
          },
          "relevance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelevanceAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "ielts_score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IELTSScore"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "pronunciation",
          "predicted_text"
        ],
        "title": "AnalyzeResponse"
      },
      "Body_create_api_key_api_admin_keys_post": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Description of the API key purpose"
          },
          "minute_limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Minute Limit",
            "description": "Maximum requests per minute",
            "default": 10
          },
          "daily_limit": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Daily Limit",
            "description": "Maximum requests per day",
            "default": 1000
          },
          "monthly_limit": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 1.0,
            "title": "Monthly Limit",
            "description": "Maximum requests per month",
            "default": 10000
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "Body_create_api_key_api_admin_keys_post"
      },
      "Body_pronunciation_analyze_pronunciation_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "expected_text": {
            "type": "string",
            "title": "Expected Text"
          }
        },
        "type": "object",
        "required": [
          "file",
          "expected_text"
        ],
        "title": "Body_pronunciation_analyze_pronunciation_post"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response"
      },
...
      "GrammarCorrection": {
        "properties": {
          "original_text": {
            "type": "string",
            "title": "Original Text"
          },
          "corrected_text": {
            "type": "string",
            "title": "Corrected Text"
          },
          "differences": {
            "items": {
              "$ref": "#/components/schemas/GrammarDifference"
            },
            "type": "array",
            "title": "Differences"
          },
          "taggedText": {
            "type": "string",
            "title": "Taggedtext"
          },
          "lexical_analysis": {
            "type": "string",
            "title": "Lexical Analysis"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "improvements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Improvements"
          },
          "lexical_band_score": {
            "type": "number",
            "title": "Lexical Band Score"
          },
          "modelAnswers": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "object",
            "title": "Modelanswers"
          },
          "grammar_score": {
            "type": "number",
            "title": "Grammar Score"
          }
        },
        "type": "object",
        "required": [
          "original_text",
          "corrected_text",
          "differences",
          "taggedText",
          "lexical_analysis",
          "strengths",
          "improvements",
          "lexical_band_score",
          "modelAnswers",
          "grammar_score"
        ],
        "title": "GrammarCorrection"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IELTSScore": {
        "properties": {
          "overall_band": {
            "type": "number",
            "title": "Overall Band"
          },
          "fluency_coherence": {
            "type": "number",
            "title": "Fluency Coherence"
          },
          "lexical_resource": {
            "type": "number",
            "title": "Lexical Resource"
          },
          "grammatical_range": {
            "type": "number",
            "title": "Grammatical Range"
          },
          "pronunciation": {
            "type": "number",
            "title": "Pronunciation"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "overall_band",
          "fluency_coherence",
          "lexical_resource",
          "grammatical_range",
          "pronunciation",
          "explanation"
        ],
        "title": "IELTSScore"
      },
...
      "PronunciationResult": {
        "properties": {
          "words": {
            "items": {
              "$ref": "#/components/schemas/WordPronunciation"
            },
            "type": "array",
            "title": "Words"
          },
          "overall_score": {
            "type": "number",
            "title": "Overall Score"
          }
        },
        "type": "object",
        "required": [
          "words",
          "overall_score"
        ],
        "title": "PronunciationResult"
      },
      "RelevanceAnalysis": {
        "properties": {
          "relevance_score": {
            "type": "number",
            "title": "Relevance Score"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "key_points_covered": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Points Covered"
          },
          "missing_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Points"
          }
        },
        "type": "object",
        "required": [
          "relevance_score",
          "explanation",
          "key_points_covered",
          "missing_points"
        ],
        "title": "RelevanceAnalysis"
      },
...
      "SpeechMetrics": {
        "properties": {
          "speech_rate": {
            "type": "number",
            "title": "Speech Rate"
          },
          "speech_rate_over_time": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Speech Rate Over Time"
          },
          "pauses": {
            "type": "integer",
            "title": "Pauses"
          },
          "filler_words": {
            "type": "integer",
            "title": "Filler Words"
          },
          "discourse_markers": {
            "items": {
              "$ref": "#/components/schemas/DiscourseMarker"
            },
            "type": "array",
            "title": "Discourse Markers"
          },
          "filler_words_per_min": {
            "type": "number",
            "title": "Filler Words Per Min"
          },
          "pause_details": {
            "items": {
              "$ref": "#/components/schemas/PauseDetail"
            },
            "type": "array",
            "title": "Pause Details"
          },
          "repetitions": {
            "items": {
              "$ref": "#/components/schemas/Repetition"
            },
            "type": "array",
            "title": "Repetitions"
          },
          "filler_words_details": {
            "items": {
              "$ref": "#/components/schemas/FillerWordDetail"
            },
            "type": "array",
            "title": "Filler Words Details"
          }
        },
        "type": "object",
        "required": [
          "speech_rate",
          "speech_rate_over_time",
          "pauses",
          "filler_words",
          "discourse_markers",
          "filler_words_per_min",
          "pause_details",
          "repetitions",
          "filler_words_details"
        ],
        "title": "SpeechMetrics"
      },
      "UsageAnalyticsResponse": {
        "properties": {
          "api_keys": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Api Keys"
          },
          "recent_logs": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Recent Logs"
          }
        },
        "type": "object",
        "required": [
          "api_keys",
          "recent_logs"
        ],
        "title": "UsageAnalyticsResponse",
        "description": "Response model for usage analytics"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API key authentication using Bearer token. Format: `Bearer sk-your-api-key-here`"
      }
    }
  }
}